ST-2[wip] Main screen imposition.   add:    - add simple logic

+++ app/src/main/java/com/nimtego/sheetamount/mvp/view_model/CalculateData.kt
@@ -3,5 +3,5 @@ package com.nimtego.sheetamount.mvp.view_model
 import java.io.Serializable
 
 data class CalculateData (
-    val current: Int
+    val current: Double
 ): Serializable
