+++ app/src/main/java/com/nimtego/sheetamount/mvp/view_model/EnterDateModel.kt
@@ -3,5 +3,5 @@ package com.nimtego.sheetamount.mvp.view_model
 import java.io.Serializable
 
 data class EnterDateModel (
-    val power: Int
+    val power: Double
 ): Serializable
